backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
beanie==1.26.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.2.0
pymongo>=4.9,<4.10
slowapi==0.1.9
python-dotenv==1.2.2
dnspython==2.7.0

